Update GeneratePlink.sh
Had some additional problems with grep matching the correct genome names. The problem was made worse with the use of underscores in the genome names. Hopefully should function as intended now.

Currently the indel and presence absence matrices don't include the reference genome even if specified in the ingroup or outgroup file. including the reference genome in the ingroup or outgroup list will cause the indel and PA matrix to be malformed. The reference can still be included as a synthetic genome and then aligned against itself.